Reorganize Gtk3Backend into a file per backend feature
