+++ Sources/Gtk3Backend/Features/Checkboxes.swift
@@ -1,0 +1,25 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.Checkboxes {
+    public func createCheckbox() -> Widget {
+        return Gtk3.CheckButton()
+    }
+
+    public func updateCheckbox(
+        _ checkboxWidget: Widget,
+        environment: EnvironmentValues,
+        onChange: @escaping (Bool) -> Void
+    ) {
+        let checkboxWidget = checkboxWidget as! Gtk3.CheckButton
+        checkboxWidget.sensitive = environment.isEnabled
+        checkboxWidget.notifyActive = { widget, _ in
+            onChange(widget.active)
+        }
+    }
+
+    public func setState(ofCheckbox checkboxWidget: Widget, to state: Bool) {
+        (checkboxWidget as! Gtk3.CheckButton).active = state
+    }
+}
