+++ Sources/Gtk3Backend/Features/FileOpenDialogs.swift
@@ -1,0 +1,23 @@
+import CGtk3
+import Foundation
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.FileOpenDialogs {
+    public func showOpenDialog(
+        fileDialogOptions: FileDialogOptions,
+        openDialogOptions: OpenDialogOptions,
+        window: Window?,
+        resultHandler handleResult: @escaping (DialogResult<[URL]>) -> Void
+    ) {
+        showFileChooserDialog(
+            fileDialogOptions: fileDialogOptions,
+            action: .open,
+            configure: { chooser in
+                chooser.selectMultiple = openDialogOptions.allowMultipleSelections
+            },
+            window: window,
+            resultHandler: handleResult
+        )
+    }
+}
