+++ Sources/Gtk3Backend/Features/IncomingURLs.swift
@@ -1,0 +1,14 @@
+import CGtk3
+import Foundation
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.IncomingURLs {
+    public func setIncomingURLHandler(to action: @escaping (URL) -> Void) {
+        gtkApp.onOpen = { urls in
+            for url in urls {
+                action(url)
+            }
+        }
+    }
+}
