+++ Sources/Gtk3Backend/Features/Paths.swift
@@ -1,0 +1,211 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.Paths {
+    public final class Path {
+        var path: SwiftCrossUI.Path?
+        var scaleFactor: Double = 1
+    }
+
+    public func createPathWidget() -> Widget {
+        Gtk3.DrawingArea()
+    }
+
+    public func createPath() -> Path {
+        Path()
+    }
+
+    public func updatePath(
+        _ path: Path,
+        _ source: SwiftCrossUI.Path,
+        bounds: SwiftCrossUI.Path.Rect,
+        pointsChanged: Bool,
+        environment: EnvironmentValues
+    ) {
+        path.path = source
+        path.scaleFactor = environment.windowScaleFactor
+    }
+
+    /// Assumes that the path backing widget has already been given the correct
+    /// size.
+    public func renderPath(
+        _ path: Path,
+        container: Widget,
+        strokeColor: SwiftCrossUI.Color.Resolved,
+        fillColor: SwiftCrossUI.Color.Resolved,
+        overrideStrokeStyle: StrokeStyle?
+    ) {
+        let drawingArea = container as! Gtk3.DrawingArea
+
+        // We don't actually care about leaking backends, but might as well use
+        // a weak reference anyway.
+        drawingArea.doDraw = { [weak self] cairo in
+            guard let self, let path = path.path else {
+                return
+            }
+
+            let fillRule: cairo_fill_rule_t
+            switch path.fillRule {
+                case .evenOdd:
+                    fillRule = CAIRO_FILL_RULE_EVEN_ODD
+                case .winding:
+                    fillRule = CAIRO_FILL_RULE_WINDING
+            }
+            cairo_set_fill_rule(cairo, fillRule)
+
+            let strokeStyle = overrideStrokeStyle ?? path.strokeStyle
+            let strokeCap: cairo_line_cap_t
+            switch strokeStyle.cap {
+                case .butt:
+                    strokeCap = CAIRO_LINE_CAP_BUTT
+                case .round:
+                    strokeCap = CAIRO_LINE_CAP_ROUND
+                case .square:
+                    strokeCap = CAIRO_LINE_CAP_SQUARE
+            }
+            cairo_set_line_cap(cairo, strokeCap)
+
+            let strokeJoin: cairo_line_join_t
+            switch strokeStyle.join {
+                case .bevel:
+                    strokeJoin = CAIRO_LINE_JOIN_BEVEL
+                case .miter(let limit):
+                    strokeJoin = CAIRO_LINE_JOIN_MITER
+                    cairo_set_miter_limit(cairo, limit)
+                case .round:
+                    strokeJoin = CAIRO_LINE_JOIN_ROUND
+            }
+            cairo_set_line_join(cairo, strokeJoin)
+
+            cairo_set_line_width(cairo, strokeStyle.width)
+
+            self.renderPathActions(path.actions, to: cairo)
+
+            let fillPattern = cairo_pattern_create_rgba(
+                Double(fillColor.red),
+                Double(fillColor.green),
+                Double(fillColor.blue),
+                Double(fillColor.opacity)
+            )
+            cairo_set_source(cairo, fillPattern)
+            cairo_fill_preserve(cairo)
+            cairo_pattern_destroy(fillPattern)
+
+            let strokePattern = cairo_pattern_create_rgba(
+                Double(strokeColor.red),
+                Double(strokeColor.green),
+                Double(strokeColor.blue),
+                Double(strokeColor.opacity)
+            )
+            cairo_set_source(cairo, strokePattern)
+            cairo_stroke(cairo)
+            cairo_pattern_destroy(strokePattern)
+        }
+
+        drawingArea.queueDraw()
+    }
+
+    private func renderPathActions(
+        _ actions: [SwiftCrossUI.Path.Action],
+        to cairo: OpaquePointer
+    ) {
+        for action in actions {
+            switch action {
+                case .transform(let transform):
+                    var matrix = cairo_matrix_t()
+                    matrix.xx = transform.linearTransform.x
+                    matrix.xy = transform.linearTransform.y
+                    matrix.yx = transform.linearTransform.z
+                    matrix.yy = transform.linearTransform.w
+                    matrix.x0 = transform.translation.x
+                    matrix.y0 = transform.translation.y
+                    cairo_transform(cairo, &matrix)
+                default:
+                    break
+            }
+        }
+
+        for (index, action) in actions.enumerated() {
+            switch action {
+                case .moveTo(let point):
+                    cairo_move_to(cairo, point.x, point.y)
+                case .lineTo(let point):
+                    if index == 0 {
+                        cairo_move_to(cairo, 0, 0)
+                    }
+                    cairo_line_to(cairo, point.x, point.y)
+                case .quadCurve(let control, let end):
+                    if index == 0 {
+                        cairo_move_to(cairo, 0, 0)
+                    }
+                    var startX = 0.0
+                    var startY = 0.0
+                    cairo_get_current_point(cairo, &startX, &startY)
+                    let start = SIMD2(startX, startY)
+                    let control1 = (start + 2 * control) / 3
+                    let control2 = (end + 2 * control) / 3
+                    cairo_curve_to(
+                        cairo,
+                        control1.x,
+                        control1.y,
+                        control2.x,
+                        control2.y,
+                        end.x,
+                        end.y
+                    )
+                case .cubicCurve(let control1, let control2, let end):
+                    if index == 0 {
+                        cairo_move_to(cairo, 0, 0)
+                    }
+                    cairo_curve_to(
+                        cairo,
+                        control1.x,
+                        control1.y,
+                        control2.x,
+                        control2.y,
+                        end.x,
+                        end.y
+                    )
+                case .rectangle(let rect):
+                    cairo_rectangle(
+                        cairo,
+                        rect.origin.x,
+                        rect.origin.y,
+                        rect.size.x,
+                        rect.size.y
+                    )
+                case .circle(let center, let radius):
+                    cairo_arc(cairo, center.x, center.y, radius, 0, 2 * .pi)
+                case .arc(
+                let center,
+                let radius,
+                let startAngle,
+                let endAngle,
+                let clockwise
+            ):
+                    let arcFunc = clockwise ? cairo_arc : cairo_arc_negative
+                    arcFunc(
+                        cairo,
+                        center.x,
+                        center.y,
+                        radius,
+                        startAngle,
+                        endAngle
+                    )
+                case .transform(let transform):
+                    var matrix = cairo_matrix_t()
+                    matrix.xx = transform.linearTransform.x
+                    matrix.xy = transform.linearTransform.y
+                    matrix.yx = transform.linearTransform.z
+                    matrix.yy = transform.linearTransform.w
+                    matrix.x0 = transform.translation.x
+                    matrix.y0 = transform.translation.y
+                    cairo_matrix_invert(&matrix)
+                    cairo_transform(cairo, &matrix)
+                case .subpath(let subpathActions):
+                    renderPathActions(subpathActions, to: cairo)
+            }
+        }
+    }
+}
