+++ Sources/Gtk3Backend/Features/ProgressBars.swift
@@ -1,0 +1,18 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.ProgressBars {
+    public func createProgressBar() -> Widget {
+        ProgressBar()
+    }
+
+    public func updateProgressBar(
+        _ widget: Widget,
+        progressFraction: Double?,
+        environment: EnvironmentValues
+    ) {
+        let progressBar = widget as! ProgressBar
+        progressBar.fraction = progressFraction ?? 0
+    }
+}
