+++ Sources/Gtk3Backend/Features/Sliders.swift
@@ -1,0 +1,33 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.Sliders {
+    public func createSlider() -> Widget {
+        let scale = Scale()
+        scale.expandHorizontally = true
+        return scale
+    }
+
+    public func updateSlider(
+        _ slider: Widget,
+        minimum: Double,
+        maximum: Double,
+        decimalPlaces: Int,
+        environment: EnvironmentValues,
+        onChange: @escaping (Double) -> Void
+    ) {
+        let slider = slider as! Scale
+        slider.sensitive = environment.isEnabled
+        slider.minimum = minimum
+        slider.maximum = maximum
+        slider.digits = decimalPlaces
+        slider.valueChanged = { widget in
+            onChange(widget.value)
+        }
+    }
+
+    public func setValue(ofSlider slider: Widget, to value: Double) {
+        (slider as! Scale).value = value
+    }
+}
