+++ Sources/Gtk3Backend/Features/StringLabelButtons.swift
@@ -1,0 +1,26 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.StringLabelButtons {
+    public func createSimpleButton() -> Widget {
+        return Button()
+    }
+
+    public func updateSimpleButton(
+        _ button: Widget,
+        label: String,
+        environment: EnvironmentValues,
+        action: @escaping () -> Void
+    ) {
+        // TODO: Update button label color using environment
+        let button = button as! Gtk3.Button
+        button.sensitive = environment.isEnabled
+        button.label = label
+        button.clicked = { _ in action() }
+        button.css.clear()
+        button.css.set(
+            properties: Self.cssProperties(for: environment, isControl: true)
+        )
+    }
+}
