+++ Sources/Gtk3Backend/Features/Switches.swift
@@ -1,0 +1,25 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.Switches {
+    public func createSwitch() -> Widget {
+        return Switch()
+    }
+
+    public func updateSwitch(
+        _ switchWidget: Widget,
+        environment: EnvironmentValues,
+        onChange: @escaping (Bool) -> Void
+    ) {
+        let switchWidget = switchWidget as! Gtk3.Switch
+        switchWidget.sensitive = environment.isEnabled
+        switchWidget.notifyActive = { widget, _ in
+            onChange(widget.active)
+        }
+    }
+
+    public func setState(ofSwitch switchWidget: Widget, to state: Bool) {
+        (switchWidget as! Gtk3.Switch).active = state
+    }
+}
