+++ Sources/Gtk3Backend/Features/TextEditors.swift
@@ -1,0 +1,36 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.TextEditors {
+    public func createTextEditor() -> Widget {
+        let textEditor = Gtk3.TextView()
+        textEditor.wrapMode = .wordCharacter
+        return textEditor
+    }
+
+    public func updateTextEditor(
+        _ textEditor: Widget,
+        environment: EnvironmentValues,
+        onChange: @escaping (String) -> Void
+    ) {
+        let textEditor = textEditor as! Gtk3.TextView
+        textEditor.buffer.changed = { buffer in
+            onChange(buffer.text)
+        }
+
+        textEditor.css.clear()
+        textEditor.css.set(properties: Self.cssProperties(for: environment, isControl: false))
+        textEditor.css.set(property: CSSProperty(key: "background", value: "none"))
+    }
+
+    public func setContent(ofTextEditor textEditor: Widget, to content: String) {
+        let textEditor = textEditor as! Gtk3.TextView
+        textEditor.buffer.text = content
+    }
+
+    public func getContent(ofTextEditor textEditor: Widget) -> String {
+        let textEditor = textEditor as! Gtk3.TextView
+        return textEditor.buffer.text
+    }
+}
