+++ Sources/Gtk3Backend/Features/TextViews.swift
@@ -1,0 +1,52 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.TextViews {
+    public func size(
+        of text: String,
+        whenDisplayedIn widget: Widget,
+        proposedWidth: Int?,
+        proposedHeight: Int?,
+        environment: EnvironmentValues
+    ) -> SIMD2<Int> {
+        let pango = Pango(for: widget)
+        let (width, height) = pango.getTextSize(
+            text,
+            ellipsize: (widget as! CustomLabel).ellipsize,
+            proposedWidth: proposedWidth.map(Double.init),
+            proposedHeight: proposedHeight.map(Double.init)
+        )
+        return SIMD2(width, height)
+    }
+
+    public func createTextView() -> Widget {
+        let textView = CustomLabel(string: "")
+        textView.horizontalAlignment = .start
+        textView.wrap = true
+        textView.lineWrapMode = .wordCharacter
+        textView.ellipsize = .end
+        return textView
+    }
+
+    public func updateTextView(
+        _ textView: Widget,
+        content: String,
+        environment: EnvironmentValues
+    ) {
+        let textView = textView as! CustomLabel
+        textView.label = content
+        textView.justify =
+            switch environment.multilineTextAlignment {
+                case .leading:
+                    Justification.left
+                case .center:
+                    Justification.center
+                case .trailing:
+                    Justification.right
+            }
+
+        textView.css.clear()
+        textView.css.set(properties: Self.cssProperties(for: environment))
+    }
+}
