+++ Sources/Gtk3Backend/Features/Widgets.swift
@@ -1,0 +1,29 @@
+import CGtk3
+import Gtk3
+@_spi(Backends) import SwiftCrossUI
+
+extension Gtk3Backend: BackendFeatures.Widgets {
+    public typealias Widget = Gtk3.Widget
+
+    public var defaultPaddingAmount: Int { 10 }
+
+    public func show(widget: Widget) {
+        widget.show()
+    }
+
+    public func tag(widget: Widget, as tag: String) {
+        widget.tag(as: tag)
+    }
+
+    public func naturalSize(of widget: Widget) -> SIMD2<Int> {
+        let currentSize = widget.getSizeRequest()
+        widget.setSizeRequest(width: -1, height: -1)
+        let (width, height) = widget.getNaturalSize()
+        widget.setSizeRequest(width: currentSize.width, height: currentSize.height)
+        return SIMD2(width, height)
+    }
+
+    public func setSize(of widget: Widget, to size: SIMD2<Int>) {
+        widget.setSizeRequest(width: size.x, height: size.y)
+    }
+}
