+++ Sources/Gtk3Backend/Helpers/TooltipContainer.swift
@@ -1,0 +1,41 @@
+import CGtk3
+import Gtk3
+
+final class TooltipContainer: Fixed {
+    private var tooltip: UnsafeMutableBufferPointer<CChar>
+
+    init(_ child: Widget) {
+        self.tooltip = UnsafeMutableBufferPointer(start: nil, count: 0)
+        super.init()
+        self.put(child, x: 0, y: 0)
+    }
+
+    deinit {
+        deallocateText()
+    }
+
+    func setTooltip(text: String) {
+        text.utf8CString.withUnsafeBufferPointer { buf in
+            // TODO(bbrk24): Should this be `>=` or `==`?
+            if tooltip.count >= buf.count {
+                strcpy(tooltip.baseAddress!, buf.baseAddress!)
+            } else {
+                deallocateText()
+
+                tooltip = .allocate(capacity: buf.count)
+                _ = tooltip.initialize(from: buf)
+            }
+        }
+
+        gtk_widget_set_tooltip_text(widgetPointer, tooltip.baseAddress)
+    }
+
+    private func deallocateText() {
+        if tooltip.count > 0 {
+            tooltip.deinitialize()
+            tooltip.deallocate()
+        }
+
+        tooltip = UnsafeMutableBufferPointer(start: nil, count: 0)
+    }
+}
